# Bump mapbox/variant to version 1.2.0