… third_party/variant/.travis.yml
… third_party/variant/CHANGELOG.md
… third_party/variant/Jamroot
… third_party/variant/Makefile
… third_party/variant/README.md
… third_party/variant/include/mapbox/variant.hpp
… third_party/variant/include/mapbox/variant_cast.hpp
… third_party/variant/include/mapbox/variant_visitor.hpp
… third_party/variant/package.json
… third_party/variant/test/compilation_failure/get_type.cpp
… third_party/variant/test/compilation_failure/mutating_visitor_on_const.cpp
… third_party/variant/test/include/catch.hpp
… third_party/variant/test/lambda_overload_test.cpp
… third_party/variant/test/recursive_wrapper_test.cpp
… third_party/variant/test/t/binary_visitor_impl.hpp
… third_party/variant/test/t/issue21.cpp
… third_party/variant/test/t/nothrow_move.cpp
… third_party/variant/test/t/optional.cpp
… third_party/variant/test/t/recursive_wrapper.cpp
… third_party/variant/test/t/sizeof.cpp
… third_party/variant/test/t/variant.cpp
… third_party/variant/test/t/variant_alternative.cpp
… third_party/variant/test/t/visitor_result_type.cpp