I hope it will allow to build it with C++20 without changes :) 

<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1103.17<br/>plain u32: 1091.43<br/>aliased double: 962.625<br/>plain double: 965.307 | aliased u32: 1100.76<br/>plain u32: 1099.72<br/>aliased double: 960.702<br/>plain double: 974.091 |
| json-render | String: 8.46336ms<br/>Stringstream: 11.1872ms<br/>Vector: 8.67628ms | String: 8.43559ms<br/>Stringstream: 11.7697ms<br/>Vector: 7.54067ms |
| match_ch | Default radius: <br/>4.49287ms/req at 82 coordinate<br/>0.0547911ms/coordinate<br/>Radius 5m: <br/>4.48831ms/req at 82 coordinate<br/>0.0547355ms/coordinate<br/>Radius 10m: <br/>15.3478ms/req at 82 coordinate<br/>0.187168ms/coordinate<br/>Radius 15m: <br/>37.3478ms/req at 82 coordinate<br/>0.455461ms/coordinate<br/>Radius 30m: <br/>316.73ms/req at 82 coordinate<br/>3.86256ms/coordinate | Default radius: <br/>4.50651ms/req at 82 coordinate<br/>0.0549575ms/coordinate<br/>Radius 5m: <br/>4.49255ms/req at 82 coordinate<br/>0.0547872ms/coordinate<br/>Radius 10m: <br/>15.3007ms/req at 82 coordinate<br/>0.186594ms/coordinate<br/>Radius 15m: <br/>37.3494ms/req at 82 coordinate<br/>0.45548ms/coordinate<br/>Radius 30m: <br/>315.538ms/req at 82 coordinate<br/>3.84802ms/coordinate |
| match_mld | Default radius: <br/>3.05762ms/req at 82 coordinate<br/>0.0372881ms/coordinate<br/>Radius 5m: <br/>2.97074ms/req at 82 coordinate<br/>0.0362285ms/coordinate<br/>Radius 10m: <br/>10.8395ms/req at 82 coordinate<br/>0.132189ms/coordinate<br/>Radius 15m: <br/>28.1103ms/req at 82 coordinate<br/>0.342808ms/coordinate<br/>Radius 30m: <br/>332.632ms/req at 82 coordinate<br/>4.05648ms/coordinate | Default radius: <br/>2.85589ms/req at 82 coordinate<br/>0.0348279ms/coordinate<br/>Radius 5m: <br/>2.84228ms/req at 82 coordinate<br/>0.034662ms/coordinate<br/>Radius 10m: <br/>10.3239ms/req at 82 coordinate<br/>0.125902ms/coordinate<br/>Radius 15m: <br/>26.3391ms/req at 82 coordinate<br/>0.321208ms/coordinate<br/>Radius 30m: <br/>300.338ms/req at 82 coordinate<br/>3.66266ms/coordinate |
| packedvector | random write:<br/>std::vector 9832.35 ms<br/>util::packed_vector 77221.2 ms<br/>slowdown: 7.85378<br/>random read:<br/>std::vector 9010.66 ms<br/>util::packed_vector 30609.8 ms<br/>slowdown: 3.39707 | random write:<br/>std::vector 9876.26 ms<br/>util::packed_vector 74274.5 ms<br/>slowdown: 7.52051<br/>random read:<br/>std::vector 8712.56 ms<br/>util::packed_vector 30409 ms<br/>slowdown: 3.49024 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>598.133ms<br/>0.598133ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>377.55ms<br/>0.37755ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>762.401ms<br/>0.762401ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>161.478ms<br/>0.161478ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>103.565ms<br/>0.103565ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>145.133ms<br/>0.145133ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>160.715ms<br/>0.160715ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>103.538ms<br/>0.103538ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>144.98ms<br/>0.14498ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>603.787ms<br/>0.603787ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>387.293ms<br/>0.387293ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>767.793ms<br/>0.767793ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>162.799ms<br/>0.162799ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>104.46ms<br/>0.10446ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>146.034ms<br/>0.146034ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>162.552ms<br/>0.162552ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>104.765ms<br/>0.104765ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>149.112ms<br/>0.149112ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>740.996ms<br/>0.740996ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>470.444ms<br/>0.470444ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>961.84ms<br/>0.96184ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>301.203ms<br/>0.301203ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>169.557ms<br/>0.169557ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>313.309ms<br/>0.313309ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>294.38ms<br/>0.29438ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>167.912ms<br/>0.167912ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>313.849ms<br/>0.313849ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>723.168ms<br/>0.723168ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>466.241ms<br/>0.466241ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>946.47ms<br/>0.94647ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>286.286ms<br/>0.286286ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>167.637ms<br/>0.167637ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>304.614ms<br/>0.304614ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>285.534ms<br/>0.285534ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>168.435ms<br/>0.168435ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>306.096ms<br/>0.306096ms/req |
| rtree | 1 result:<br/>207.085ms ->  0.0207085 ms/query<br/>10 results:<br/>242.61ms ->  0.024261 ms/query | 1 result:<br/>207.727ms ->  0.0207727 ms/query<br/>10 results:<br/>241.787ms ->  0.0241787 ms/query |
<!-- BENCHMARK_RESULTS_END -->