@SiarheiFedartsou 2024-05-24
> :+1: 
> 
> 
> 
> Eventually, we should aim replacing this with `std::variant`.

Sure, it is the plan, but unfortunately it is not a drop-in replacement.